+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Impl `Eq` for `Config`;
 - `Config` builder;
 - Impl `FromStr` for `Config`;
-- Adds `Config::from_env`;
+- `Config::from_env`;
+- `inspect::databases()`;
 
 ## [5.2.0] - 2026-04-28